+++ packages/adaptive_navigation/lib/src/adaptive_navigation_scaffold.dart
@@ -33,10 +33,9 @@ class AdaptiveScaffoldDestination {
   final IconData icon;
 
   const AdaptiveScaffoldDestination({
-    @required this.title,
-    @required this.icon,
-  })  : assert(title != null),
-        assert(icon != null);
+    required this.title,
+    required this.icon,
+  });
 }
 
 /// A widget that adapts to the current display size, displaying a [Drawer],
@@ -44,9 +43,9 @@ class AdaptiveScaffoldDestination {
 /// defined in the [destinations] parameter.
 class AdaptiveNavigationScaffold extends StatelessWidget {
   const AdaptiveNavigationScaffold({
-    Key key,
+    Key? key,
     this.appBar,
-    this.body,
+    required this.body,
     this.floatingActionButton,
     this.floatingActionButtonLocation,
     this.floatingActionButtonAnimator,
@@ -63,53 +62,51 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
     this.drawerEdgeDragWidth,
     this.drawerEnableOpenDragGesture = true,
     this.endDrawerEnableOpenDragGesture = true,
-    @required this.selectedIndex,
-    @required this.destinations,
+    required this.selectedIndex,
+    required this.destinations,
     this.onDestinationSelected,
     this.navigationTypeResolver,
     this.drawerHeader,
     this.fabInRail = true,
     this.includeBaseDestinationsInMenu = true,
-  })  : assert(selectedIndex != null),
-        assert(destinations != null),
-        super(key: key);
+  })  : super(key: key);
 
   /// See [Scaffold.appBar].
-  final Widget appBar;
+  final PreferredSizeWidget? appBar;
 
   /// See [Scaffold.body].
   final Widget body;
 
   /// See [Scaffold.floatingActionButton].
-  final FloatingActionButton floatingActionButton;
+  final FloatingActionButton? floatingActionButton;
 
   /// See [Scaffold.floatingActionButtonLocation].
   ///
   /// Ignored if [fabInRail] is true.
-  final FloatingActionButtonLocation floatingActionButtonLocation;
+  final FloatingActionButtonLocation? floatingActionButtonLocation;
 
   /// See [Scaffold.floatingActionButtonAnimator].
   ///
   /// Ignored if [fabInRail] is true.
-  final FloatingActionButtonAnimator floatingActionButtonAnimator;
+  final FloatingActionButtonAnimator? floatingActionButtonAnimator;
 
   /// See [Scaffold.persistentFooterButtons].
-  final List<Widget> persistentFooterButtons;
+  final List<Widget>? persistentFooterButtons;
 
   /// See [Scaffold.endDrawer].
-  final Widget endDrawer;
+  final Widget? endDrawer;
 
   /// See [Scaffold.drawerScrimColor].
-  final Color drawerScrimColor;
+  final Color? drawerScrimColor;
 
   /// See [Scaffold.backgroundColor].
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// See [Scaffold.bottomSheet].
-  final Widget bottomSheet;
+  final Widget? bottomSheet;
 
   /// See [Scaffold.resizeToAvoidBottomInset].
-  final bool resizeToAvoidBottomInset;
+  final bool? resizeToAvoidBottomInset;
 
   /// See [Scaffold.primary].
   final bool primary;
@@ -124,7 +121,7 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
   final bool extendBodyBehindAppBar;
 
   /// See [Scaffold.drawerEdgeDragWidth].
-  final double drawerEdgeDragWidth;
+  final double? drawerEdgeDragWidth;
 
   /// See [Scaffold.drawerEnableOpenDragGesture].
   final bool drawerEnableOpenDragGesture;
@@ -148,15 +145,15 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
   /// The stateful widget that creates the adaptive scaffold needs to keep
   /// track of the index of the selected [AdaptiveScaffoldDestination] and call
   /// `setState` to rebuild the adaptive scaffold with the new [selectedIndex].
-  final ValueChanged<int> onDestinationSelected;
+  final ValueChanged<int>? onDestinationSelected;
 
   /// Determines the navigation type that the scaffold uses.
-  final NavigationTypeResolver navigationTypeResolver;
+  final NavigationTypeResolver? navigationTypeResolver;
 
   /// The leading item in the drawer when the navigation has a drawer.
   ///
   /// If null, then there is no header.
-  final Widget drawerHeader;
+  final Widget? drawerHeader;
 
   /// Whether the [floatingActionButton] is inside or the rail or in the regular
   /// spot.
@@ -183,15 +180,14 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
     return Drawer(
       child: ListView(
         children: [
-          if (drawerHeader != null) drawerHeader,
+          // TODO: Find a better way to write `drawerHeader!`
+          if (drawerHeader != null) drawerHeader!,
           for (int i = 0; i < destinations.length; i++)
             ListTile(
               leading: Icon(destinations[i].icon),
               title: Text(destinations[i].title),
               onTap: () {
-                if (onDestinationSelected != null) {
-                  onDestinationSelected(i);
-                }
+                onDestinationSelected?.call(i);
               },
             )
         ],
@@ -208,7 +204,7 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
     final drawerDestinations = destinations.length > bottomNavigationOverflow
         ? destinations.sublist(
             includeBaseDestinationsInMenu ? 0 : bottomNavigationOverflow)
-        : [];
+        : <AdaptiveScaffoldDestination>[];
     return Scaffold(
       body: body,
       appBar: appBar,
@@ -240,7 +236,7 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
     final drawerDestinations = destinations.length > railDestinationsOverflow
         ? destinations.sublist(
             includeBaseDestinationsInMenu ? 0 : railDestinationsOverflow)
-        : [];
+        : <AdaptiveScaffoldDestination>[];
     return Scaffold(
       appBar: appBar,
       drawer: drawerDestinations.isEmpty
@@ -295,7 +291,8 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
       drawer: Drawer(
         child: Column(
           children: [
-            if (drawerHeader != null) drawerHeader,
+            // TODO: Find a better way to write `drawerHeader!`
+            if (drawerHeader != null) drawerHeader!,
             for (final destination in destinations)
               ListTile(
                 leading: Icon(destination.icon),
@@ -331,7 +328,8 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
         Drawer(
           child: Column(
             children: [
-              if (drawerHeader != null) drawerHeader,
+              // TODO: Find a better way to write `drawerHeader!`
+              if (drawerHeader != null) drawerHeader!,
               for (final destination in destinations)
                 ListTile(
                   leading: Icon(destination.icon),
@@ -387,13 +385,12 @@ class AdaptiveNavigationScaffold extends StatelessWidget {
       case NavigationType.permanentDrawer:
         return _buildPermanentDrawerScaffold();
     }
-    throw Exception('Invalid navigation type: $navigationType');
   }
 
   void _destinationTapped(AdaptiveScaffoldDestination destination) {
     final index = destinations.indexOf(destination);
     if (index != selectedIndex) {
-      onDestinationSelected(index);
+      onDestinationSelected?.call(index);
     }
   }
 }