+++ packages/adaptive_breakpoints/CHANGELOG.md
@@ -1,10 +1,14 @@
+# [0.0.4] - December 14, 2020
+
+Migrate to NNBD
+
 # [0.0.3] - September 21, 2020
 
-Removed adaptive container widget and updated adaptive window types enums.
+Remove adaptive container widget and update adaptive window types enums.
 
 # [0.0.2] - August 24, 2020
 
-Added operator logic to AdaptiveWindowType.
+Add operator logic to AdaptiveWindowType.
 
 # [0.0.1] - August 12, 2020