+++ packages/adaptive_navigation/CHANGELOG.md
@@ -1,3 +1,7 @@
+# [0.0.3] - December 14, 2020
+
+Migrate to NNBD
+
 # [0.0.2] - October 9, 2020
 
 Update adaptive_breakpoint dependency.
@@ -5,4 +9,3 @@ Update adaptive_breakpoint dependency.
 # [0.0.1] - August 12, 2020
 
 The initial release adds [adaptive navigation](https://material.io/design/layout/responsive-layout-grid.html#ui-regions)
-